+++ Requests.jsx
@@ -1,20 +1,41 @@
-import React from 'react';
-import { Grid } from "@mui/material";
-import { Header } from "./components/header.jsx";
-import { useState , useRef, useEffect } from 'react';
-import { Accordion, AccordionSummary, AccordionDetails, Box, Typography, TextField, Button, RadioGroup, FormControlLabel, Radio } from '@mui/material';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import Paper from '@mui/material/Paper';
+import { React, useState , useRef, useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { Grid, Accordion, AccordionSummary, AccordionDetails, Box, Typography, TextField, Button, RadioGroup, FormControlLabel, Radio, Paper, IconButton } from '@mui/material';
 import { DataGrid } from '@mui/x-data-grid';
-import IconButton from '@mui/material/IconButton';
-import ArticleIcon from '@mui/icons-material/Article';
+import { Header } from "./components/header.jsx";
 import { RequestDetails } from './components/request-details.jsx';
+import { ApprobeRequest } from './components/approve-request.jsx';
+import { RejectRequest } from './components/reject-request.jsx';
+import { format } from 'date-fns';
+import Cookies from "universal-cookie";
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
+import ArticleIcon from '@mui/icons-material/Article';
+import CheckIcon from '@mui/icons-material/Check';
+import CloseIcon from '@mui/icons-material/Close';
 import CheckCircleOutlineIcon from '@mui/icons-material/CheckCircleOutline';
 import PauseCircleOutlineIcon from '@mui/icons-material/PauseCircleOutline';
 import HighlightOffIcon from '@mui/icons-material/HighlightOff';
 
-export const RequestsUser = () => {
-    const [value, setValue] = React.useState('todas');
+export const Requests = () => {
+    const cookies = new Cookies();
+    const navigate = useNavigate();
+
+    let token = cookies.get("token");
+    let rol = cookies.get("rol");
+
+    if (!token) {
+        navigate('/');
+    }
+
+    const headers = new Headers();
+    headers.append("Authorization", `${token}`);
+
+    const options = {
+        method: "GET",
+        headers: headers
+    };
+
+    const [radioStatus, setRadioStatus] = useState('ALL');
 
     const clearFields = (refs) => {
         refs.forEach((ref) => {
@@ -26,35 +47,43 @@ export const RequestsUser = () => {
         });
     };
 
-    const textNombre = useRef();
+    const textSolicitante = useRef();
     const textEmail = useRef();
     const textEstacion = useRef();
     const dateSolicitud = useRef();
 
     const handleClearFields = () => {
-        clearFields([textNombre, textEmail, textEstacion, dateSolicitud]);
+        clearFields([textSolicitante, textEmail, textEstacion, dateSolicitud]);
+        fetchData();
     };
 
     const handleChange = (event) => {
-        setValue(event.target.value);
+        setRadioStatus(event.target.value);
     };
 
     const handleBuscarClick = () => {
-        const nombre = textNombre.current.value;
-        const fechaSolicitud = dateSolicitud.current.value;
-        const email = textEmail.current.value;
-        const estacion = textEstacion.current.value;
-      
+        let created_by = textSolicitante.current.value;
+        let created_at = dateSolicitud.current.value;
+        let email = textEmail.current.value;
+        let name = textEstacion.current.value;
+        let status = '';
+
+        if (radioStatus !== 'ALL') {
+            status = radioStatus;
+        }
+
         const filteredRows = rows.filter(row => {
-          return row.created_by === nombre &&
-                 row.created_at === fechaSolicitud &&
-                 row.name === estacion &&
-                 row.email === email;
+          return (
+                (created_by === '' || row.created_by === created_by) &&
+                (created_at === '' || row.created_at.slice(0, 10).includes(created_at)) &&
+                (name === '' || row.name.toUpperCase().includes(name.toUpperCase())) &&
+                (email === '' || row.email.toUpperCase().includes(email.toUpperCase()) &&
+                (status === '' || row.status === status))
+            );
         });
 
         setRows(filteredRows);
-    };
-
+    }
 
     const getStatusIcon = (status) => {
         if (status === 'APPROVED') {
@@ -79,21 +108,38 @@ export const RequestsUser = () => {
         setOpenDetails(false);
     };
 
-    const token = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJyZXNwaXJBUiIsImlhdCI6MTY4MzE1NjQzMSwiZXhwIjoxNzQ2MjI4NDMxLCJhdWQiOiJ3d3cuZXhhbXBsZS5jb20iLCJzdWIiOiJqcm9ja2V0QGV4YW1wbGUuY29tIiwiaWQiOiIxIiwidXNlcm5hbWUiOiJKb2huRG9lIiwicm9sIjoidXNlciJ9.RPweKkWBmGOMR0z8-HLMgH-6dAzmGBak2dBoSGbCS9U";
-    
-    const headers = new Headers();
-    headers.append("Authorization", `${token}`);
+    const [openApprove, setOpenApprove] = useState(false);
 
-    const options = {
-        method: "GET",
-        headers: headers
+    const handleOpenApprove = (row) => {
+        setSelectedRow(row);
+        setOpenApprove(true);
     };
 
+    const handleCloseApprove = () => {
+        setOpenApprove(false);
+    };
+
+    const [openReject, setOpenReject] = useState(false);
+
+    const handleOpenReject = (row) => {
+        setSelectedRow(row);
+        setOpenReject(true);
+    };
+
+    const handleCloseReject = () => {
+        setOpenReject(false);
+    };
+
+    function formatDate(dateString) {
+        const date = new Date(dateString);
+        return format(date, 'MM/dd/yyyy');
+    }
+
     const columns = [
-        { field: 'id', headerName: 'ID', width: 70 },
+        { field: 'id', headerName: 'Id', width: 70 },
         { field: 'created_by', headerName: 'Solicitante', width: 150 },
         { field: 'email', headerName: 'Correo Electrónico', width: 170 },
-        { field: 'created_at', headerName: 'Fecha de Solicitud', width: 130 },
+        { field: 'created_at', headerName: 'Fecha de Solicitud', width: 130, valueFormatter: (params) => formatDate(params.value) },
         { field: 'name', headerName: 'Estación', width: 130 },
         { field: 'serial_number', headerName: 'Nº de Serie', width: 130 },
         {
@@ -115,11 +161,23 @@ export const RequestsUser = () => {
             width: 150,
             disableColumnMenu: true,
             renderCell: (params) => {
+                const isInactive = params.row.status === 'APPROVED' || params.row.status === 'REJECTED';
+
                 return (
                     <>
                         <IconButton onClick={() => handleOpenDetails(params.row)}>
                             <ArticleIcon />
                         </IconButton>
+                        {rol === "admin" && (
+                            <>
+                                <IconButton onClick={() => handleOpenApprove(params.row)} disabled={isInactive}>
+                                    <CheckIcon />
+                                </IconButton>
+                                <IconButton onClick={() => handleOpenReject(params.row)} disabled={isInactive}>
+                                    <CloseIcon />
+                                </IconButton>
+                            </>
+                        )}
                     </>
                 );
             },
@@ -129,6 +187,30 @@ export const RequestsUser = () => {
     const [rows, setRows] = useState([]);
 
     useEffect(() => {
+        fetch("http://localhost:8080/requests?pageSize=0&page=0", options)
+        .then(response => response.json())
+        .then((data) =>
+            setRows(
+              data.map((item) => ({
+                id: item._id,
+                serial_number: item.serial_number,
+                name: item.name,
+                longitude: item.longitude,
+                latitude: item.latitude,
+                brand: item.brand,
+                model: item.model,
+                status: item.status,
+                created_by: item.created_by,
+                created_at: item.created_at,
+                approved_by: item.approved_by,
+                approved_at: item.approved_at
+              }))
+            )
+        )
+        .catch(error => console.error(error));
+    }, []);
+
+    const fetchData = () => {
         fetch("http://localhost:8080/requests?pageSize=0&page=0", options)
         .then(response => response.json())
         .then((data) =>
@@ -150,7 +232,7 @@ export const RequestsUser = () => {
             )
         )
         .catch(error => console.error(error));
-    }, []);
+    };
 
     return(
         <Grid container direction="column">
@@ -182,10 +264,10 @@ export const RequestsUser = () => {
                                         <AccordionDetails sx={{ display: 'flex', justifyContent: 'center', flexDirection: 'column', width: '80%', margin: 'auto' }}>
                                             <Grid container spacing={2}>
                                                 <Grid item xs={12} sm={6}>
-                                                    <TextField label="Nombre" inputRef={textNombre} fullWidth />
+                                                    <TextField label="Solicitante" inputRef={textSolicitante} fullWidth />
                                                 </Grid>
                                                 <Grid item xs={12} sm={6}>
-                                                    <TextField label="Fecha de solicitud" inputRef={dateSolicitud} fullWidth type="date" InputLabelProps={{ shrink: true }} />
+                                                    <TextField label="Fecha de Solicitud" inputRef={dateSolicitud} fullWidth type="date" InputLabelProps={{ shrink: true }} />
                                                 </Grid>
                                                 <Grid item xs={12} sm={6}>
                                                     <TextField label="Correo Electrónico" inputRef={textEmail} fullWidth type="email"/>
@@ -195,13 +277,13 @@ export const RequestsUser = () => {
                                                 </Grid>
                                                 <Grid item xs={12} sm={6}>
                                                     <Box sx={{ display: 'flex', alignItems: 'center', flexWrap: 'wrap'}}>
-                                                    <Typography variant="body1" sx={{ mr: 1 }} >Estado:</Typography>
-                                                    <RadioGroup value={value} onChange={handleChange} row sx={{ alignItems: 'center' }}>
-                                                        <FormControlLabel value="ALL" control={<Radio />} label="Todas" />
-                                                        <FormControlLabel value="APPROVED" control={<Radio />} label="Aprobadas" />
-                                                        <FormControlLabel value="REJECTED" control={<Radio />} label="Rechazadas" />
-                                                        <FormControlLabel value="PENDING" control={<Radio />} label="Pendientes" />
-                                                    </RadioGroup>
+                                                        <Typography variant="body1" sx={{ mr: 1 }} >Estado:</Typography>
+                                                        <RadioGroup value={radioStatus} onChange={handleChange} row sx={{ alignItems: 'center' }}>
+                                                            <FormControlLabel value="ALL" control={<Radio />} label="Todas" />
+                                                            <FormControlLabel value="APPROVED" control={<Radio />} label="Aprobadas" />
+                                                            <FormControlLabel value="REJECTED" control={<Radio />} label="Rechazadas" />
+                                                            <FormControlLabel value="PENDING" control={<Radio />} label="Pendientes" />
+                                                        </RadioGroup>
                                                     </Box>
                                                 </Grid>
                                                 <Grid item xs={12} sm={6}>
@@ -260,8 +342,7 @@ export const RequestsUser = () => {
                                                     disableColumnFilter
                                                     columnVisibilityModel={{
                                                         id: false,
-                                                    }}
-                                                    disableColumnSelector
+                                                    }} 
                                                     style={{ overflowX: 'auto', backgroundColor: '#A9B4C4'}}
                                                 />
                                             </div>
@@ -281,6 +362,24 @@ export const RequestsUser = () => {
                     open={openDetails}
                     onClose={handleCloseDetails}
                     rowData={selectedRow}
+                />
+            )}
+
+            {selectedRow && (
+                <ApprobeRequest
+                    open={openApprove}
+                    onClose={handleCloseApprove}
+                    rowData={selectedRow}
+                    TransitionProps={{ timeout: 100 }}
+                />
+            )}
+
+            {selectedRow && (
+                <RejectRequest
+                    open={openReject}
+                    onClose={handleCloseReject}
+                    rowData={selectedRow}
+                    TransitionProps={{ timeout: 100 }}
                 />
             )}
         </Grid>
