+++ components/edit-station.jsx
@@ -1,21 +1,19 @@
-import React from 'react';
+import { React, useState } from "react";
+import { useNavigate } from 'react-router-dom';
 import { TextField, Button, Dialog, DialogActions, DialogContent, DialogTitle, Grid } from '@mui/material' ;
-import { useState } from "react";
-import { styled } from '@mui/system';
-
-const Form = styled('form')({
-  display: 'flex',
-  flexDirection: 'column',
-  margin: theme => theme.spacing(2)
-});
-
-const Field = styled(TextField)({
-  margin: theme => theme.spacing(1),
-  width: '250px',
-  padding: '8px'
-});
+import Cookies from "universal-cookie";
 
 export const EditStation = ({ open, onClose, rowData }) =>{
+  const cookies = new Cookies();
+  const navigate = useNavigate();
+
+  let token = cookies.get("token");
+  let id_user = cookies.get("id_user");
+
+  if (!token) {
+      navigate('/');
+  }
+
   const handleClose = () => {
       onClose();
   };
@@ -24,12 +22,8 @@ export const EditStation = ({ open, onClose, rowData }) =>{
     name: rowData.name
   });
 
-  const token = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJyZXNwaXJBUiIsImlhdCI6MTY4MzE1NjQzMSwiZXhwIjoxNzQ2MjI4NDMxLCJhdWQiOiJ3d3cuZXhhbXBsZS5jb20iLCJzdWIiOiJqcm9ja2V0QGV4YW1wbGUuY29tIiwiaWQiOiIxIiwidXNlcm5hbWUiOiJKb2huRG9lIiwicm9sIjoiYWRtaW4ifQ.4AdK8vzb0ec-m6jjGp8aLFoO4Prn6fFwjJmeqiwBS8s";
-
   const handleEdit = async () => {
     try {
-      const userId = 1;
-
       const data = {
         name: formData.name
       };
@@ -43,7 +37,7 @@ export const EditStation = ({ open, onClose, rowData }) =>{
         body: JSON.stringify(data)
       };
   
-      const response = await fetch(`http://localhost:8080/users/${userId}/stations/${rowData.id}/rename`, options);
+      await fetch(`http://localhost:8080/users/${id_user}/stations/${rowData.id}/rename`, options);
   
       handleClose();
       window.location.reload();
@@ -56,26 +50,26 @@ export const EditStation = ({ open, onClose, rowData }) =>{
     <Dialog open={open} onClose={handleClose} aria-labelledby="form-dialog-title">
       <DialogTitle id="form-dialog-title">Editar Estación</DialogTitle>
       <DialogContent dividers>
-        <Form>
-            <Grid container spacing={2}>
-                <Grid item xs={6}>
-                    <Field
-                        required
-                        autoFocus
-                        margin="dense"
-                        id="name"
-                        label="Nombre"
-                        variant="outlined"
-                        value={formData.name}
-                        onChange={(e) =>
-                          setFormData((prevFormData) => ({
-                            ...prevFormData,
-                            name: e.target.value,
-                          }))
-                        }
-                    />
-                </Grid>
-            </Grid>
+          <Grid container spacing={2}>
+              <Grid item xs={6}>
+                  <TextField
+                      required
+                      autoFocus
+                      margin="dense"
+                      id="name"
+                      label="Nombre"
+                      variant="outlined"
+                      style = {{width: 250}}
+                      value={formData.name}
+                      onChange={(e) =>
+                        setFormData((prevFormData) => ({
+                          ...prevFormData,
+                          name: e.target.value,
+                        }))
+                      }
+                  />
+              </Grid>
+          </Grid>
           <DialogActions>
             <Button onClick={handleClose} color="primary">
               Cancelar
@@ -84,7 +78,6 @@ export const EditStation = ({ open, onClose, rowData }) =>{
               Guardar
             </Button>
           </DialogActions>
-        </Form>
       </DialogContent>
     </Dialog>
   );
