+++ components/reject-request.jsx
@@ -1,26 +1,47 @@
-import React from 'react';
-import { Button, Dialog, DialogActions, DialogContent, DialogTitle, Typography } from '@mui/material' ;
+import { React, useState } from "react";
+import { Button, Dialog, DialogActions, DialogContent, DialogTitle, Typography, Grid, TextField } from '@mui/material';
+import { useNavigate } from 'react-router-dom';
+import Cookies from "universal-cookie";
 
 export const RejectRequest = ({ open, onClose, rowData }) => {
+    const cookies = new Cookies();
+    const navigate = useNavigate();
+
+    let token = cookies.get("token");
+    let rol = cookies.get("rol");
+
+    if (!token && rol !== "admin") {
+        navigate('/');
+    }
+
     const handleClose = () => {
         onClose();
     };
 
-    const token = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJyZXNwaXJBUiIsImlhdCI6MTY4MzE1NjQzMSwiZXhwIjoxNzQ2MjI4NDMxLCJhdWQiOiJ3d3cuZXhhbXBsZS5jb20iLCJzdWIiOiJqcm9ja2V0QGV4YW1wbGUuY29tIiwiaWQiOiIxIiwidXNlcm5hbWUiOiJKb2huRG9lIiwicm9sIjoiYWRtaW4ifQ.4AdK8vzb0ec-m6jjGp8aLFoO4Prn6fFwjJmeqiwBS8s";
+    const [formData, setFormData] = useState({
+        reason: ''
+    });
     
-    const headers = new Headers();
-    headers.append("Authorization", `${token}`);
-
-    const options = {
-        method: "PATCH",
-        headers: headers
-    };
-    
-    const handleApprove = async () => {
+    const handleReject = async () => {
         try {
-           await fetch(`http://localhost:8080/users/${rowData.created_by}/requests/${rowData.id}/reject`, options);
-           handleClose();
-           window.location.reload();
+            debugger;
+
+            const data = {
+                reason: formData.reason
+            };
+
+            const options = {
+                method: "PATCH",
+                headers: {
+                    'Authorization': `${token}`,
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify(data)
+            };
+
+            await fetch(`http://localhost:8080/users/${rowData.created_by}/requests/${rowData.id}/reject`, options);
+            handleClose();
+            window.location.reload();
         } catch (error) {
           console.log(error);
         }
@@ -32,12 +53,32 @@ export const RejectRequest = ({ open, onClose, rowData }) => {
             <DialogContent>
                 <DialogContent dividers>
                     <Typography align="center">¿Esta seguro de rechazar la Estación {rowData.station} con Nº de Serie {rowData.serial_number}?<br/>Se enviaria un Email de aviso a la dirección {rowData.email}</Typography>
+                    <Grid container alignItems="center" justifyContent="center">
+                        <Grid item>
+                            <TextField
+                                required
+                                autoFocus
+                                margin="dense"
+                                id="reason"
+                                label="Razón"
+                                variant="outlined"
+                                style = {{width: 250}}
+                                value={formData.reason}
+                                onChange={(e) =>
+                                setFormData((prevFormData) => ({
+                                    ...prevFormData,
+                                    reason: e.target.value,
+                                }))
+                                }
+                            />
+                        </Grid>
+                    </Grid>
                 </DialogContent>
                 <DialogActions>
                     <Button onClick={handleClose} color="primary">
                     Cancelar
                     </Button>
-                    <Button onClick={handleApprove} color="primary">
+                    <Button onClick={handleReject} color="primary">
                     Rechazar
                     </Button>
                 </DialogActions>
