+++ components/request-details.jsx
@@ -1,11 +1,21 @@
-import React from 'react';
+import { React } from 'react';
+import { useNavigate } from 'react-router-dom';
 import { Grid, Dialog, DialogContent, DialogTitle, Typography } from '@mui/material' ;
+import { format } from 'date-fns';
+import Cookies from "universal-cookie";
 import IconButton from '@mui/material/IconButton';
 import CloseIcon from '@mui/icons-material/Close';
-import { format } from 'date-fns';
-
 
 export const RequestDetails = ({ open, onClose, rowData }) =>{
+    const cookies = new Cookies();
+    const navigate = useNavigate();
+
+    let token = cookies.get("token");
+    
+    if (!token) {
+        navigate('/');
+    }
+
     const handleClose = () => {
         onClose();
     };
@@ -72,7 +82,7 @@ export const RequestDetails = ({ open, onClose, rowData }) =>{
                     </Grid>
                     <Grid item xs={5}>
                         <Typography>Longitud</Typography>
-                        <Typography variant="h6">{rowData.longitud}</Typography>
+                        <Typography variant="h6">{rowData.longitude}</Typography>
                     </Grid>
                     <Grid item xs={5}>
                         <Typography>Latitud</Typography>
