+++ package.json
@@ -13,7 +13,8 @@
     "@emotion/styled": "^11.10.8",
     "@expo/webpack-config": "^18.0.1",
     "@mui/icons-material": "^5.11.16",
-    "@mui/material": "^5.12.3",
+    "@mui/material": "^5.13.0",
+    "@mui/styled-engine-sc": "^5.12.0",
     "@mui/x-data-grid": "^6.3.1",
     "@mui/x-date-pickers": "^6.3.1",
     "axios": "^1.4.0",
@@ -29,12 +30,14 @@
     "jsrsasign": "^10.8.6",
     "jwa": "^2.0.0",
     "jws": "^4.0.0",
+    "jwt-decode": "^3.1.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-native": "0.71.7",
     "react-native-web": "~0.18.10",
     "react-router-dom": "^6.11.1",
     "stream": "^0.0.2",
+    "styled-components": "^5.3.10",
     "universal-cookie": "^4.0.4",
     "util": "^0.12.5"
   },
