--- styles/login.css
@@ -1,129 +1,0 @@
-.logo1 {
-  position: relative;
-}
-.login-logo-section {
-  position: absolute;
-  top: 0;
-  left: calc(50% - 829px);
-  width: 1658px;
-  height: 240px;
-  overflow: hidden;
-  display: flex;
-  flex-direction: row;
-  padding: 10px;
-  box-sizing: border-box;
-  align-items: flex-end;
-  justify-content: center;
-}
-.iniciar-sesion {
-  position: relative;
-  font-weight: 500;
-}
-.email-input {
-  align-self: stretch;
-  border-radius: 4px;
-  border: 1px solid #79747e;
-}
-.iniciar-sesion1 {
-  position: relative;
-  letter-spacing: 0.1px;
-  line-height: 20px;
-  font-weight: 500;
-}
-.login-button {
-  border-radius: 100px;
-  background-color: #7ea2d9;
-  box-shadow: 0 4px 4px rgba(0, 0, 0, 0.25);
-  width: 123px;
-  height: 38px;
-  flex-shrink: 0;
-  overflow: hidden;
-  display: flex;
-  flex-direction: column;
-  padding: 10px 24px;
-  box-sizing: border-box;
-  align-items: center;
-  justify-content: center;
-}
-.login-button-group,
-.login-form {
-  background-color: #d7e1f0;
-  display: flex;
-  align-items: center;
-}
-.login-button-group {
-  overflow: hidden;
-  flex-direction: row;
-  padding: 5px;
-  justify-content: center;
-  font-size: 14px;
-  color: #fff;
-}
-.login-form {
-  border-radius: 16px;
-  border: 1px solid #cac4d0;
-  box-sizing: border-box;
-  width: 354px;
-  height: 304px;
-  flex-shrink: 0;
-  flex-direction: column;
-  padding: 20px;
-  justify-content: flex-start;
-  gap: 20px;
-}
-.login-container,
-.login-section {
-  overflow: hidden;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.login-container {
-  width: 363px;
-  height: 309px;
-  flex-shrink: 0;
-  flex-direction: column;
-}
-.login-section {
-  position: absolute;
-  top: 240px;
-  left: calc(50% - 829px);
-  width: 1658px;
-  height: 347px;
-  flex-direction: row;
-  padding: 10px;
-  box-sizing: border-box;
-  font-size: 36px;
-}
-.login-dark-mode-icon {
-  position: relative;
-  width: 45px;
-  height: 45px;
-  flex-shrink: 0;
-  overflow: hidden;
-}
-.login-bottom-section {
-  position: absolute;
-  top: 587px;
-  left: calc(50% - 829px);
-  width: 1658px;
-  height: 287px;
-  overflow: hidden;
-  display: flex;
-  flex-direction: row;
-  padding: 10px;
-  box-sizing: border-box;
-  align-items: flex-end;
-  justify-content: flex-end;
-}
-.login {
-  position: relative;
-  background-color: #fff;
-  width: 100%;
-  height: 874px;
-  overflow: hidden;
-  text-align: center;
-  font-size: 80px;
-  color: #000;
-  font-family: Product Sans Regular;
-}
